requirements.txt
streamlit
tensorflow-cpu
numpy==1.26.4
pandas
scikit-learn==1.1.3
pickle5
